% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{add_debugger_file}
\alias{add_debugger_file}
\title{add_debugger_file}
\usage{
add_debugger_file(path_from_file, input_dir, output_dir)
}
\arguments{
\item{path_from_file}{R file which you want to update}

\item{input_dir}{Path to the shiny application}

\item{output_dir}{Path to output directory where you want to save shiny application . default value is NULL to overwrite.}
}
\description{
This function takes R file and update it with message
}
